CV calculation:         fixed the bug that would result in empty spots on CV map         fixed the bug that would crop the map for 200 by 200 recordings; Condition parameters:         fixed the bug that would not allow for concave polygons as masks; apdMap:         fixed the bug that would crop the map for 200 by 200 recordings; ensembleAverage:         fixed the bug that would break the averaging for non-round FPS; interface:         disabled popup toolbars on movie screens 	fixed the bug that would make file selection box invisible 	fixed the bug that would make rhythm not recognize .bin files CMOSconverter: 	fixed the bug that would set acqFreq to the same value as FPS in newly converted .bin files

+++ apdMap.m
@@ -55,7 +55,7 @@ APD_max_rescaled = maxapd * Fs / 1000;
 
 AP_level = 1.0 - percentAPD / 100;
 
-area_coords = int8(area_coords);
+area_coords = int16(area_coords);
 j_min = 1 + area_coords(1);
 i_min = 1 + area_coords(2);
 j_max = area_coords(1) + area_coords(3);
